refactor: 이동과 관련된 파일을 별도의 패키지로 구조화
