+++ src/main/java/domain/move/movement/OrthogonalThenDiagonalMovement.java
@@ -1,4 +1,4 @@
-package domain.piece.movement;
+package domain.move.movement;
 
 import domain.board.Intersection;
 import domain.direction.Direction;
