+++ src/main/java/domain/move/moverule/BasicMoveRule.java
@@ -1,9 +1,9 @@
-package domain.piece.moverule;
+package domain.move.moverule;
 
 import domain.board.Intersection;
 import domain.game.Side;
 import domain.piece.AlivePieces;
-import domain.piece.movement.Path;
+import domain.move.movement.Path;
 import java.util.List;
 
 public final class BasicMoveRule {
