+++ src/test/java/domain/move/movement/SingleStepMovementTest.java
@@ -1,12 +1,10 @@
-package domain.piece.movement;
+package domain.move.movement;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.board.Intersection;
 import domain.game.Side;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -16,26 +14,25 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-@DisplayName("직선 이동 테스트")
-class StraightLineMovementTest {
+@DisplayName("모든 방향 1칸 이동 테스트")
+class SingleStepMovementTest {
 
-    private StraightLineMovement movement;
+    private SingleStepMovement movement;
 
     @BeforeEach
     void setUp() {
-        movement = new StraightLineMovement();
+        movement = new SingleStepMovement();
     }
 
-    @DisplayName("앞, 뒤, 양 옆으로 칸 수 제약없이 이동이 가능하다")
+    @DisplayName("앞, 뒤, 양 옆으로 1칸 이동이 가능하다")
     @Nested
-    class 모든_빙향으로_칸_수_제약없이_이동이_가능하다 {
+    class 모든_빙향으로_1칸_이동이_가능하다 {
 
         @DisplayName("진영에 상관없이 동일하게 적용된다")
         @ParameterizedTest
         @EnumSource(Side.class)
         void 진영에_상관없이_동일하게_적용된다(Side side) {
             Intersection currentIntersection = new Intersection(5, 5);
-            List<Intersection> expected = allDirectionIntersectionsExcludeCurrentIntersection(currentIntersection);
 
             List<Path> movablePaths = movement.movablePaths(currentIntersection, side);
             List<Intersection> movableDestinations = movablePaths.stream()
@@ -43,7 +40,12 @@ class StraightLineMovementTest {
                     .toList();
 
             assertThat(movableDestinations)
-                    .containsExactlyInAnyOrderElementsOf(expected);
+                    .containsExactlyInAnyOrder(
+                            new Intersection(4, 5),
+                            new Intersection(6, 5),
+                            new Intersection(5, 4),
+                            new Intersection(5, 6)
+                    );
         }
     }
 
@@ -73,35 +75,14 @@ class StraightLineMovementTest {
         private static Stream<Arguments> allBorderIntersections() {
             return Stream.of(
                     Arguments.of(new Intersection(1, 1),
-                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(1, 1))),
+                            List.of(new Intersection(1, 2), new Intersection(2, 1))),
                     Arguments.of(new Intersection(1, 9),
-                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(1, 9))),
+                            List.of(new Intersection(1, 8), new Intersection(2, 9))),
                     Arguments.of(new Intersection(10, 1),
-                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(10, 1))),
+                            List.of(new Intersection(9, 1), new Intersection(10, 2))),
                     Arguments.of(new Intersection(10, 9),
-                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(10, 9)))
+                            List.of(new Intersection(9, 9), new Intersection(10, 8)))
             );
         }
-    }
-
-    private static List<Intersection> allDirectionIntersectionsExcludeCurrentIntersection(
-            Intersection currentIntersection
-    ) {
-        int currentRow = currentIntersection.row();
-        int currentFile = currentIntersection.file();
-
-        List<Intersection> expected = new ArrayList<>();
-
-        IntStream.rangeClosed(1, 10)
-                .filter(row -> row != currentRow)
-                .mapToObj(row -> new Intersection(row, currentFile))
-                .forEach(expected::add);
-
-        IntStream.rangeClosed(1, 9)
-                .filter(file -> file != currentFile)
-                .mapToObj(file -> new Intersection(currentRow, file))
-                .forEach(expected::add);
-
-        return expected;
     }
 }
